\docType{package}
\name{GOglm-package}
\alias{GOglm}
\alias{GOglm-package}
\title{Length Bias Correction in Gene Ontology Enrichment Analysis Using
Logistic Regression}
\description{
  This GOglm package is a beta version under development.
  The goglm function implements the GOglm method discussed
  in Mi et al. (PLOS ONE, 7(10): e46128).  The package
  includes a summarized RNA-Seq data example (the prostate
  cancer dataset) for methodological illustrations.
}
\author{
  Gu Mi <mig@stat.oregonstate.edu>, Yanming Di
  <diy@stat.oregonstate.edu>. Maintainer: Gu Mi
  <http://people.oregonstate.edu/~mig>
}
\references{
  Mi G, Di Y, Emerson S, Cumbie JS and Chang JH (2012)
  "Length bias correction in Gene Ontology enrichment
  analysis using logistic regression", 7(10): e46128.
}
\keyword{package}

