\name{[.prepGOglm}
\alias{[.prepGOglm}
\title{Subset a data frame of class "prepGOglm"}
\usage{
  \method{[}{prepGOglm} (x, ..., drop = TRUE)
}
\arguments{
  \item{x}{a data frame with class \code{prepGOglm}}

  \item{...}{for future use}

  \item{drop}{logical value}
}
\value{
  A subset of \code{x}.
}
\description{
  This is an S3 generic function for \code{[} with the
  class \code{prepGOglm}.
}
\author{
  Gu Mi \email{mig@stat.oregonstate.edu}, Yanming Di
  \email{diy@stat.oregonstate.edu}
}

