\docType{package}
\name{GOglm-package}
\alias{GOglm}
\alias{GOglm-package}
\title{Length Bias Correction in Gene Ontology Enrichment Analysis Using
Logistic Regression}
\description{
  This GOglm package is a beta version under development.
  The goglm function implements the GOglm method discussed
  in Mi et al. (PLOS ONE, in press).  The package includes
  a summarized RNA-Seq data example (the prostate cancer
  dataset) for methodological illustrations.
}
\details{
  \tabular{ll}{ Package: \tab GOglm\cr Type: \tab
  Package\cr Version: \tab 0.2.0\cr Depends: \tab R (>=
  2.10)\cr Imports: \tab goseq\cr LazyData: \tab yes\cr
  LazyLoad: \tab yes\cr License: \tab GPL-2\cr URL: \tab
  http://people.oregonstate.edu/~mig\cr Collate: \tab
  'goglm.R' 'package.R' 'plot.R' 'plot.prepGOglm.R'
  'prepare.R' 'revMap.R' 'summary.R' 'summary.goglm.R'
  'summary.prepGOglm.R' 'unfactor.R' 'ProsCan_DE.R'
  'ProsCan_Length.R'\cr Built: \tab R 2.15.1; ; 2012-09-26
  06:44:36 UTC; unix\cr }
}
\author{
  Gu Mi <mig@stat.oregonstate.edu>, Yanming Di
  <diy@stat.oregonstate.edu>.

  Maintainer: Gu Mi <mig@stat.oregonstate.edu>
}
\references{
  Mi G, Di Y, Emerson S, Cumbie JS and Chang JH (2012)
  "Length bias correction in Gene Ontology enrichment
  analysis using logistic regression", 7(10): e46128.
}
\keyword{package}

